Finalizado o desktop de cursos/ Sem definição do tamanho
ACABOU!
