+++ app/src/diario/Cursos/nbproject/configs/Executar_como_WebStart.properties
@@ -1,0 +1,2 @@
+# Do not modify this property in this configuration. It can be re-generated.
+$label=Executar como WebStart
