+++ app/src/diario/Cursos/nbproject/configs/Executar_no_Browser.properties
@@ -1,0 +1,2 @@
+# Do not modify this property in this configuration. It can be re-generated.
+$label=Executar no Browser
