+++ app/src/diario/Cursos/nbproject/configs/Run_as_WebStart.properties
@@ -1,0 +1,2 @@
+# Do not modify this property in this configuration. It can be re-generated.
+$label=Run as WebStart
