--- app/src/diário/Cursos/nbproject/configs/Executar_como_WebStart.properties
@@ -1,2 +1,0 @@
-# Do not modify this property in this configuration. It can be re-generated.
-$label=Executar como WebStart
