--- app/src/diário/Cursos/src/Principal/FXML.fxml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.text.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" onMouseEntered="#refresh" prefHeight="914.0" prefWidth="1310.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Principal.TableController">
-   <children>
-      <TableView fx:id="table" layoutX="185.0" layoutY="205.0" mouseTransparent="true" prefHeight="689.0" prefWidth="940.0" style="-fx-border-color: #ffffff;" stylesheets="@style.css" AnchorPane.bottomAnchor="20.0" AnchorPane.leftAnchor="185.0" AnchorPane.rightAnchor="185.0" AnchorPane.topAnchor="205.0">
-         <columns>
-            <TableColumn fx:id="col_id" prefWidth="75.0" styleClass="coluna" text="ID" />
-            <TableColumn fx:id="col_idDepto" prefWidth="75.0" styleClass="coluna" text="ID-DEPTO" />
-            <TableColumn fx:id="col_nome" prefWidth="209.0" styleClass="coluna" text="NOME" />
-            <TableColumn fx:id="col_horas" prefWidth="133.0" styleClass="coluna" text="HORAS-TOTAL" />
-            <TableColumn fx:id="col_modalidade" prefWidth="130.0" styleClass="coluna" text="MODALIDADE" />
-            <TableColumn fx:id="col_funcoes" prefWidth="115.0" styleClass="coluna" text="AÇÕES" />
-            <TableColumn fx:id="col_funcoes1" prefWidth="115.0" styleClass="coluna" />
-            <TableColumn fx:id="col_funcoes2" prefWidth="115.0" styleClass="coluna" />
-         </columns>
-      </TableView>
-      <Label prefHeight="51.0" prefWidth="1310.0" style="-fx-background-color: #29A58C;" text="S.A." textFill="WHITE">
-         <font>
-            <Font size="35.0" />
-         </font>
-      </Label>
-      <Button fx:id="adiciona" layoutX="1046.0" layoutY="180.0" mnemonicParsing="false" onAction="#Insere" style="-fx-background-color: #29A58C;" text="ADICIONAR" textFill="WHITE" />
-      <Label layoutX="89.0" layoutY="86.0" text="Cursos">
-         <font>
-            <Font size="65.0" />
-         </font>
-      </Label>
-   </children>
-</AnchorPane>
