--- app/src/diário/Cursos/src/trabalhofinaldesktop/FXML.fxml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="596.0" prefWidth="1006.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="trabalhofinaldesktop.FXMLController">
-   <children>
-      <TableView fx:id="table" layoutX="8.0" layoutY="17.0" prefHeight="596.0" prefWidth="1062.0" style="-fx-border-color: #ffffff;" stylesheets="@style.css" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-         <columns>
-            <TableColumn fx:id="col_id" prefWidth="75.0" text="id" />
-            <TableColumn fx:id="col_idDepto" prefWidth="75.0" text="id-depto" />
-            <TableColumn fx:id="col_nome" prefWidth="209.0" text="nome" />
-            <TableColumn fx:id="col_horas" prefWidth="85.0" text="horas-total" />
-            <TableColumn fx:id="col_modalidade" prefWidth="178.0" text="modalidade" />
-            <TableColumn fx:id="col_funcoes" prefWidth="115.0" text="funções" />
-            <TableColumn fx:id="col_funcoes1" prefWidth="115.0" />
-            <TableColumn fx:id="col_funcoes2" prefWidth="115.0" />
-         </columns>
-      </TableView>
-   </children>
-   <padding>
-      <Insets bottom="10.0" left="30.0" right="10.0" top="10.0" />
-   </padding>
-</AnchorPane>
