--- app/src/diário/Cursos/src/trabalhofinaldesktop/Main.java
@@ -1,27 +1,0 @@
-
-package trabalhofinaldesktop;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-public class Main extends Application {
-    
-    @Override
-    public void start(Stage stage) throws Exception {
-	Parent root = FXMLLoader.load(getClass().getResource("FXML.fxml"));
-	
-	Scene scene = new Scene(root);
-	
-	stage.setScene(scene);
-	stage.show();
-    }
-
-    
-    public static void main(String[] args) {
-	launch(args);
-    }
-    
-}
